fix: return explicit error if no tags are found
`DescribeTag` has inconsistent behaviour if a pattern is supplied.
If no pattern is supplied, not finding any tags is not an error, whereas
if a pattern is supplied, it is an error to not find any tags.

Remove the ambiguity by returning an explicit error describing what
occurred and let `main` decide how to handle it.