feat: add support for calculating the next prerelease tag
Adds a `prerelease` command which will increment the prerelease version
based on what the next version will be from the git log, taking into
account the current prerelease version from the latest tag.

If `--pre-release` is passed in and contains a number then this will be
used and all calculations will be skipped. If `--pre-release` is passed
in and it matches the pre-release suffix of the current tag then
calculations will still occur.

Example:
```bash
$ svu current
v1.2.3-alpha.2+123

$ svu prerelease
v1.2.3-alpha.3

$ svu prerelease --pre-release alpha.33 --build 243
v1.2.3-alpha.33+243
```

Signed-off-by: Bradley Jones <jones.bradley@me.com>